+++ src/main/java/com/nusinfineon/util/OutputAnalysisCalculation.java
@@ -1,0 +1,287 @@
+package com.nusinfineon.util;
+
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
+
+/**
+ * This class provides the core logic and code that goes into processing the output excel file and generating relevant
+ * summary statistics.
+ */
+public class OutputAnalysisCalculation {
+
+    /**
+     * Calculates the overall worth of products that have passed through.
+     *
+     * Logic: From the Daily Throughput sheet, looks at the product key and quantity out columns.
+     * Sums up the overall product quantity outputted and multiplies it with the product's corresponding
+     * cost in the product-cost table, a locally stored excel file.
+     *
+     * @param dailyThroughputSheet Daily Throughput sheet.
+     * @param productCostSheet Product cost sheet.
+     * @return
+     */
+    public static HashMap<String, Double> calculateTotalProductWorth(Sheet dailyThroughputSheet, Sheet productCostSheet) {
+
+        // Read from product-key-cost table and store data
+        HashMap<String, Double> mapOfProductToCost = new HashMap<String, Double>();
+        for (int rowIndex = 1; rowIndex < productCostSheet.getPhysicalNumberOfRows(); rowIndex ++) {
+            Row row = productCostSheet.getRow(rowIndex);
+            String productKey = row.getCell(1).getStringCellValue();
+            Double productCost = row.getCell(2).getNumericCellValue();
+            mapOfProductToCost.put(productKey, productCost);
+        }
+
+        // Get Counts of Product that has been processed / outputted
+        final String QTY_OUT_COLUMN = "Qty Out";
+        final String PRODUCT_COLUMN = "Product";
+
+        // Get column index for QTY_OUT_COLUMN index
+        int qtyOutColumnIndex = -1; int productColumnIndex = -1;
+        Row headerRow = dailyThroughputSheet.getRow(0);
+        for (int cellIndex = 0; cellIndex < headerRow.getPhysicalNumberOfCells(); cellIndex ++) {
+            String cellValue = headerRow.getCell(cellIndex).getStringCellValue();
+            switch (cellValue) {
+                case QTY_OUT_COLUMN:
+                    qtyOutColumnIndex = cellIndex;
+                    break;
+                case PRODUCT_COLUMN:
+                    productColumnIndex = cellIndex;
+                    break;
+                default:
+                    break;
+            } // End of switch case block
+        } // End of for loop block
+
+        // Get unique counts for each product
+        HashMap<String, Double> mapOfProductOutCounts = new HashMap<String, Double>();
+        for (int rowIndex = 1; rowIndex < dailyThroughputSheet.getPhysicalNumberOfRows(); rowIndex++) {
+            Row row = dailyThroughputSheet.getRow(rowIndex);
+            if (row.getCell(productColumnIndex) != null) {
+                String productKey = row.getCell(productColumnIndex).getStringCellValue().trim();
+                Double productQty = row.getCell(qtyOutColumnIndex).getNumericCellValue();
+
+                if (mapOfProductOutCounts.containsKey(productKey)) {
+                    // Update entry
+                    Double currentQty = mapOfProductOutCounts.get(productKey);
+                    mapOfProductOutCounts.put(productKey, currentQty + productQty);
+                } else {
+                    // Create a new entry
+                    mapOfProductOutCounts.put(productKey, productQty);
+                }
+            }
+        } // End of fot loop
+
+        // Get the total worth of products outputted.
+        Double totalWorth = 0.0;
+        for (String productKey : mapOfProductOutCounts.keySet()) {
+            Double productCost = 0.0;
+            // If an associated cost exists
+            if (mapOfProductToCost.containsKey(productKey)) {
+                productCost = mapOfProductToCost.get(productKey);
+            } else {
+                productCost = OutputAnalysisCalculation.getAverageCostOfProductHashMap(mapOfProductToCost);
+            }
+
+            Double productWorth = productCost * mapOfProductOutCounts.get(productKey);
+            totalWorth += productWorth;
+        }
+
+        HashMap<String, Double> totalWorthMap = new HashMap<String, Double>();
+        totalWorthMap.put("TOTAL WORTH", totalWorth);
+
+        return totalWorthMap;
+    }
+
+    public static HashMap<String, Double> calculateAverageProductCycleTime(Sheet productCycleTimeSheet) {
+        final String STAYTIME_AVERAGE_COLUMN = "StayTime Average (hr)";
+        final String STAYTIME_MIN_COLUMN = "StayTime Min (hr)";
+        final String STAYTIME_MAX_COLUMN = "StayTime Max (hr)";
+
+        // Get index of column names and initialize counts
+        HashMap<String, Integer> mapOfStayTimeToIndex = new HashMap<String, Integer>();
+        HashMap<String, Long> mapOfStayTimeToCounts = new HashMap<String, Long>();
+        Row headerRow = productCycleTimeSheet.getRow(0);
+        for (int cellIndex = 0; cellIndex < headerRow.getPhysicalNumberOfCells(); cellIndex ++) {
+            String cellValue = headerRow.getCell(cellIndex).getStringCellValue();
+            switch (cellValue) {
+                case STAYTIME_AVERAGE_COLUMN:
+                    mapOfStayTimeToIndex.put(STAYTIME_AVERAGE_COLUMN, cellIndex);
+                    mapOfStayTimeToCounts.put(STAYTIME_AVERAGE_COLUMN, Long.valueOf(0));
+                    break;
+                case STAYTIME_MIN_COLUMN:
+                    mapOfStayTimeToIndex.put(STAYTIME_MIN_COLUMN, cellIndex);
+                    mapOfStayTimeToCounts.put(STAYTIME_MIN_COLUMN, Long.valueOf(0));
+                    break;
+                case STAYTIME_MAX_COLUMN:
+                    mapOfStayTimeToIndex.put(STAYTIME_MAX_COLUMN, cellIndex);
+                    mapOfStayTimeToCounts.put(STAYTIME_MAX_COLUMN, Long.valueOf(0));
+                    break;
+                default:
+                    break;
+            } // End of switch case block
+        } // End of for loop block
+
+        // Iterate through all rows and get the counts
+        for (int rowIndex = 1; rowIndex < productCycleTimeSheet.getPhysicalNumberOfRows(); rowIndex ++) {
+            Row currentRow = productCycleTimeSheet.getRow(rowIndex);
+
+            // Populate counts
+            for (String columnName: mapOfStayTimeToIndex.keySet()) {
+                int columnIndex = mapOfStayTimeToIndex.get(columnName);
+                long cellValue = (long) (currentRow.getCell(columnIndex).getNumericCellValue());
+                long currentCount = mapOfStayTimeToCounts.get(columnName);
+                mapOfStayTimeToCounts.put(columnName, currentCount + cellValue);
+            }
+        }
+
+        // Average out the Stay-Cycle Time
+        HashMap<String, Double> mapOfStayTimeToAverageStayTime = new HashMap<String, Double>();
+        for (String cycleTime: mapOfStayTimeToCounts.keySet()) {
+            long currentValue = mapOfStayTimeToCounts.get(cycleTime);
+            Double averageCycleTime = (double) currentValue / (double) (productCycleTimeSheet.getPhysicalNumberOfRows() - 1);
+            mapOfStayTimeToAverageStayTime.put(cycleTime, averageCycleTime);
+        }
+
+        return mapOfStayTimeToAverageStayTime;
+    }
+
+    /**
+     * Obtains a hash map of Average Utilization Rates of IBIS Oven.
+     *
+     *  Logic: Extract Column Headers first. Then, filter by IBIS rows and sum up the cells corresponding to each
+     *  header. Lastly, average out the numbers and only extract "averaged values" > 0.
+     *
+     * @param utilizationSheet Excel sheet that has the IBIS Utilization rates.
+     * @return A hash map of utilization categories to their respective rates.
+     * @throws IOException
+     */
+    public static HashMap<String, Double> calculateAverageIbisOvenUtilRate(Sheet utilizationSheet) {
+
+        // Get column names and corresponding index. Assumes first 2 columns are "platform" and "name"
+        HashMap<Integer, String> mapOfIndexToColumnName = new HashMap<Integer, String>();
+        HashMap<String, Double> mapOfColumnNameToTotalUtilRate = new HashMap<String, Double>();
+        Row headerRow = utilizationSheet.getRow(0);
+        for (int cellIndex = 2; cellIndex < headerRow.getPhysicalNumberOfCells(); cellIndex++) {
+            String columnName = headerRow.getCell(cellIndex).getStringCellValue();
+            mapOfIndexToColumnName.put(cellIndex, columnName);
+            mapOfColumnNameToTotalUtilRate.put(columnName, 0.0);
+        }
+
+        // Get all rows regarding IBIS.
+        ArrayList<Row> ibisRows = new ArrayList<Row>();
+        for (int rowIndex = 1; rowIndex < utilizationSheet.getPhysicalNumberOfRows(); rowIndex++) {
+            Row currentRow = utilizationSheet.getRow(rowIndex);
+            String platformType = currentRow.getCell(0).getStringCellValue();
+
+            if (platformType.equals("IBIS")) {
+                ibisRows.add(currentRow);
+            }
+        }
+
+        // Get the average utilization rate for each column
+        for (Row ibisRow : ibisRows) {
+            // Increment the counts for each row.
+            for (Integer index: mapOfIndexToColumnName.keySet()) {
+                String columnName = mapOfIndexToColumnName.get(index);
+                Double cellValue = ibisRow.getCell(index).getNumericCellValue();
+                Double currentUtilRate = mapOfColumnNameToTotalUtilRate.get(columnName);
+                mapOfColumnNameToTotalUtilRate.put(columnName, currentUtilRate + cellValue);
+            }
+        }
+        HashMap<String, Double> mapOfColumnNameToAveragelUtilRate = new HashMap<String, Double>();
+        for (String columnName: mapOfColumnNameToTotalUtilRate.keySet()) {
+            Double rateSum = mapOfColumnNameToTotalUtilRate.get(columnName);
+            if (rateSum > 0) { // Remove 0 entries
+                Double averageRate = rateSum / ibisRows.size();
+                mapOfColumnNameToAveragelUtilRate.put(columnName, averageRate);
+            }
+        }
+
+        return mapOfColumnNameToAveragelUtilRate;
+    }
+
+    /**
+     *
+     *  Logic: Looks at the input and output for the whole factory.
+     *  Input = "Load/Burn In/Transfer Normal Qty" + "Load/Burn In/Transfer YRTP Qty"
+     *  Output = "Unload Normal Qty" + "Unload YRTP Qty"
+     *
+     * @param dailyThroughputSheet
+     * @return
+     */
+    public static HashMap<String, Double> calculateThroughputBasedOnDailyThroughput(Sheet dailyThroughputSheet) {
+
+        final String LOAD_NORMAL_COLUMN = "Load/Burn In/Transfer Normal Qty";
+        final String LOAD_YRTP_COLUMN = "Load/Burn In/Transfer YRTP Qty";
+        final String UNLOAD_NORMAL_COLUMN = "Unload Normal Qty";
+        final String UNLOAD_YRTP_COLUMN = "Unload YRTP Qty";
+
+        // Get index of column names and initialize counts
+        HashMap<String, Integer> mapOfThroughputToIndex = new HashMap<String, Integer>();
+        HashMap<String, Long> mapOfThroughputToCounts = new HashMap<String, Long>();
+        Row headerRow = dailyThroughputSheet.getRow(0);
+        for (int cellIndex = 0; cellIndex < headerRow.getPhysicalNumberOfCells(); cellIndex ++) {
+            String cellValue = headerRow.getCell(cellIndex).getStringCellValue();
+            switch (cellValue) {
+                case LOAD_NORMAL_COLUMN:
+                    mapOfThroughputToIndex.put(LOAD_NORMAL_COLUMN, cellIndex);
+                    mapOfThroughputToCounts.put(LOAD_NORMAL_COLUMN, Long.valueOf(0));
+                    break;
+                case LOAD_YRTP_COLUMN:
+                    mapOfThroughputToIndex.put(LOAD_YRTP_COLUMN, cellIndex);
+                    mapOfThroughputToCounts.put(LOAD_YRTP_COLUMN, Long.valueOf(0));
+                    break;
+                case UNLOAD_NORMAL_COLUMN:
+                    mapOfThroughputToIndex.put(UNLOAD_NORMAL_COLUMN, cellIndex);
+                    mapOfThroughputToCounts.put(UNLOAD_NORMAL_COLUMN, Long.valueOf(0));
+                    break;
+                case UNLOAD_YRTP_COLUMN:
+                    mapOfThroughputToIndex.put(UNLOAD_YRTP_COLUMN, cellIndex);
+                    mapOfThroughputToCounts.put(UNLOAD_YRTP_COLUMN, Long.valueOf(0));
+                    break;
+                default:
+                    break;
+            } // End of switch case block
+        } // End of for loop block
+
+        // Iterate through all rows and get the counts
+        for (int rowIndex = 1; rowIndex < dailyThroughputSheet.getPhysicalNumberOfRows(); rowIndex ++) {
+            Row currentRow = dailyThroughputSheet.getRow(rowIndex);
+
+            // Populate counts
+            for (String columnName: mapOfThroughputToIndex.keySet()) {
+                int columnIndex = mapOfThroughputToIndex.get(columnName);
+                long cellValue = (long) (currentRow.getCell(columnIndex).getNumericCellValue());
+                long currentCount = mapOfThroughputToCounts.get(columnName);
+                mapOfThroughputToCounts.put(columnName, currentCount + cellValue);
+            }
+        }
+
+        // Sum up to input and output
+        HashMap<String, Double> mapOfInputAndOutput = new HashMap<String, Double>();
+        mapOfInputAndOutput.put("TOTAL INPUT", (double) mapOfThroughputToCounts.get(LOAD_NORMAL_COLUMN)
+                + mapOfThroughputToCounts.get(LOAD_YRTP_COLUMN));
+        mapOfInputAndOutput.put("TOTAL OUTPUT", (double) mapOfThroughputToCounts.get(UNLOAD_NORMAL_COLUMN)
+                + mapOfThroughputToCounts.get(UNLOAD_YRTP_COLUMN));
+
+        return mapOfInputAndOutput;
+    }
+
+    /**
+     * Calculates the average cost (value) of a String-Double hash map.
+     * @param mapOfProductToCost Hash map of product to cost.
+     * @return Double.
+     */
+    public static Double getAverageCostOfProductHashMap(HashMap<String, Double> mapOfProductToCost) {
+        Double sum = 0.0;
+        for (String product: mapOfProductToCost.keySet()) {
+            sum += mapOfProductToCost.get(product);
+        }
+        return (sum / mapOfProductToCost.size());
+    }
+}
+
